requirements.txt
opencv-python
pygame
numpy
pyaudio
python-dotenv
google-genai
pvporcupine
sounddevice
soundfile
